+++ apk/Changelog.md
@@ -9,6 +9,18 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Web Version May 08 2023
+==============
+
+* Add Top VOD section back
+* Fix Top Clip and VOD
+* This section's API is undocumented and a little unsealable, but works most of the time, if it fails refresh it
+
+Web Version May 05 2023
+==============
+
+* Fix VOD chat
+
 Web Version February 25 2023
 ==============